Move reg_key into separate file
